+ Implement ServiceManager
The ServiceManager class keeps track of registered micro-services, starts/stops
them on demand, and allows services to request access to each other.

A reference count is kept for each service, and when the usage goes to 0, the
service is stopped. This prevents background services from keeping connections
open, that may not be shareable (such as for video streaming)