Add initial support for the AnkerMake M5C
This change allows to monitor and send MQTT messages from/to the
AnkerMake M5C on the command line as well as observe the printer status
via the Web GUI.
It is also possible to send print jobs from your slicer to the printer
without using AnkerMake Slicer or Studio (see README.md).
Due to the M5's additional hardware (camera, light) still being expected
to exist, the Web GUI will continuously wait for a camera connection.
This should be fixed in the future.